feat: use index alias in Elasticsearch
Use an alias instead of directly pointing to a particular index.
This is helpful if the existing mapping is incompatible with the data.
Since ES doesn't allow modifications to already indexed fields
(this would make existing docs using the old index incompatible),
a new index with the correct mapping must be made.
Once the new index is created and the data are reindexed to it,
the alias is switched it. From INCA's POV, nothing needs to change
since it always refers to the alias and not to an actual index.