feat: add standardize URL processor
The scraped documents already have a `standardized_url` field which comes from an older version of urlExpander. Due to an update to the `standardize_url()`  logic, this processor will be used to re-run `standardize_url()` on the `resolved_url` key. The new version should be output to a new `standardized_url_2` field.