feat: add boolean processors
Basic processors for labeling boolean fields. These processors label all documents which are passed to it with either `True` or `False`, so they should be used when you want to add a new boolean key using custom logic that isn't available in INCA and/or doesn't match INCA's design for processors.

In INCA, the  processor class typically takes one field as input, processes it, and outputs the result to a new field. A processor can take `extra_fields` as well, as long as they already exist on the document. These processors make INCA flexible for creating new boolean fields.