fix/style:  fix TypeError for `response_code`
Previously assumed that all response codes are numeric. This commit handles cases when the response code is not numeric or can't be cast to an int. Assign these cases with `-1` (to match the ES mapping for `response_code` which is numeric).

Error message:
`TypeError: ufunc 'isnan' not supported for the input types, and the inputs could not be safely coerced to any supported types according to the casting rule ''safe''`